requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile C:\Users\aless\Desktop\Claude-Fisio\agentic-RAG\fisio-rag+graph\requirements.in -o C:\Users\aless\Desktop\Claude-Fisio\agentic-RAG\fisio-rag+graph\requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via
    #   fsspec
    #   instructor
    #   langchain-community
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   pytest-anyio
    #   starlette
appdirs==1.4.4
    # via ragas
asyncpg==0.30.0
    # via -r requirements.in
attrs==25.3.0
    # via
    #   aiohttp
    #   outcome
    #   trio
bcrypt==4.3.0
    # via passlib
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via trio
charset-normalizer==3.3.2
    # via requests
click==8.2.1
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
dataclasses-json==0.6.7
    # via langchain-community
datasets==4.0.0
    # via ragas
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
diskcache==5.6.3
    # via
    #   graphiti-core
    #   instructor
distro==1.9.0
    # via openai
docstring-parser==0.16
    # via instructor
fastapi==0.115.13
    # via -r requirements.in
filelock==3.18.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.6.0
    # via
    #   datasets
    #   huggingface-hub
    #   torch
graphiti-core==0.12.4
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   langsmith
    #   openai
httpx-sse==0.4.0
    # via langchain-community
huggingface-hub==0.34.2
    # via
    #   datasets
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   trio
    #   yarl
iniconfig==2.0.0
    # via pytest
instructor==1.10.0
    # via -r requirements.in
jinja2==3.1.4
    # via
    #   instructor
    #   torch
jiter==0.10.0
    # via
    #   instructor
    #   openai
joblib==1.5.1
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.27
    # via
    #   langchain-community
    #   ragas
langchain-community==0.3.27
    # via ragas
langchain-core==0.3.72
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
    #   ragas
langchain-openai==0.3.28
    # via ragas
langchain-text-splitters==0.3.9
    # via langchain
langsmith==0.4.8
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
lxml==6.0.0
    # via python-docx
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
mypy-extensions==1.1.0
    # via typing-inspect
neo4j==5.28.1
    # via
    #   -r requirements.in
    #   graphiti-core
nest-asyncio==1.6.0
    # via ragas
networkx==3.3
    # via torch
numpy==1.26.4
    # via
    #   datasets
    #   graphiti-core
    #   langchain-community
    #   pandas
    #   pgvector
    #   pyarrow
    #   ragas
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   transformers
openai==1.90.0
    # via
    #   graphiti-core
    #   instructor
    #   langchain-openai
    #   ragas
orjson==3.11.1
    # via langsmith
outcome==1.3.0.post0
    # via trio
packaging==25.0
    # via
    #   datasets
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   pytest
    #   transformers
pandas==2.3.1
    # via datasets
passlib==1.7.4
    # via -r requirements.in
pgvector==0.4.1
    # via -r requirements.in
pillow==11.3.0
    # via sentence-transformers
pluggy==1.6.0
    # via pytest
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
psutil==5.9.8
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pyarrow==16.1.0
    # via datasets
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   fastapi
    #   graphiti-core
    #   instructor
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via
    #   instructor
    #   pydantic
pydantic-settings==2.10.0
    # via langchain-community
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyjwt==2.8.0
    # via -r requirements.in
pymupdf==1.24.1
    # via -r requirements.in
pymupdfb==1.24.1
    # via pymupdf
pysbd==0.3.4
    # via ragas
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-anyio
pytest-anyio==0.0.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-docx==1.1.2
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   graphiti-core
    #   pydantic-settings
pytz==2024.1
    # via
    #   neo4j
    #   pandas
pyyaml==6.0.2
    # via
    #   datasets
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
    #   transformers
ragas==0.1.10
    # via -r requirements.in
regex==2024.11.6
    # via
    #   tiktoken
    #   transformers
requests==2.32.3
    # via
    #   datasets
    #   fsspec
    #   huggingface-hub
    #   instructor
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
    #   transformers
requests-toolbelt==1.0.0
    # via langsmith
rich==14.0.0
    # via
    #   instructor
    #   typer
safetensors==0.5.3
    # via transformers
scikit-learn==1.7.1
    # via sentence-transformers
scipy==1.14.0
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==3.0.1
    # via -r requirements.in
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
    #   trio
sortedcontainers==2.4.0
    # via trio
sqlalchemy==2.0.31
    # via
    #   langchain
    #   langchain-community
starlette==0.46.2
    # via fastapi
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via
    #   graphiti-core
    #   instructor
    #   langchain-community
    #   langchain-core
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.9.0
    # via
    #   langchain-openai
    #   ragas
tokenizers==0.19.1
    # via transformers
torch==2.7.1
    # via sentence-transformers
tqdm==4.67.1
    # via
    #   datasets
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.42.3
    # via
    #   -r requirements.in
    #   sentence-transformers
trio==0.30.0
    # via -r requirements.in
typer==0.16.0
    # via instructor
typing-extensions==4.14.0
    # via
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   sqlalchemy
    #   torch
    #   typer
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn==0.34.3
    # via -r requirements.in
xxhash==3.5.0
    # via datasets
yarl==1.20.1
    # via aiohttp
zstandard==0.23.0
    # via langsmith
